+++ website/public/blog/qwenpaw-developer-day-collection.en.md
@@ -1,0 +1,148 @@
+---
+title: "QwenPaw Developer Day Collection"
+date: 2026-07-01
+author: QwenPaw Team
+tags: [QwenPaw, developer-day]
+cover: /blog/qwenpaw-developer-day-collection-cover.png
+excerpt: "Replay archive from QwenPaw developer day sessions — in-depth technical talks and practical insights for every QwenPaw developer and enthusiast."
+---
+
+Last updated June 30, 2026
+
+---
+
+**06-30 QwenPaw Developer Day: QwenPaw 2.0 Driver Deep Dive**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643334363335313437375f323034353035363233375f30
+
+**06-25 QwenPaw Developer Day: AgentScope X OceanBase Live Stream**
+
+Meeting link: https://weixin.qq.com/sph/AO91dHFgqM
+
+**06-23 QwenPaw Developer Day: QwenPaw Console Updates & User Message Queue**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643333373935343937395f323034353035363233375f30
+
+**06-16 QwenPaw Developer Day: Plugin Spotlight — PawGit & TeamChat**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643333313236353132335f323034353035363233375f30
+
+**06-11 QwenPaw Developer Day: QwenPaw Plugin Marketplace**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643332353835303430355f323034353035363233375f30
+
+**06-09 QwenPaw Developer Day: QwenPaw Coding Mode**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643332323838383539345f323034353035363233375f30
+
+**06-04 QwenPaw Developer Day: ReMe Upgrade**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643331373333363833335f323034353035363233375f30
+
+**06-02 QwenPaw Developer Day: QwenPaw + Rokid Glass = ? — "Third Eye" Ant Hackathon Finalist**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643331343533333032325f323034353035363233375f30
+
+**05-28 QwenPaw Developer Day: AgentScope 2.0 Deep Dive**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643330383738303630325f323034353035363233375f30
+
+**05-26 QwenPaw Developer Day: Official Plugin CloudPaw Deep Dive**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643330353931383433345f323034353035363233375f30
+
+**05-21 QwenPaw Developer Day: Cron Jobs & Inbox Upgrade**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643330303635303130395f323034353035363233375f30
+
+**05-19 QwenPaw Developer Day: Best Practice — Building a Personal Workstation with QwenPaw**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643239373239363837315f323034353035363233375f30
+
+**05-14 QwenPaw Developer Day: QwenPaw Plugin System Upgrade**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643239313839373636305f323034353035363233375f30
+
+**05-12 QwenPaw Developer Day: QwenPaw Doctor Overview**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643238393133373436365f323034353035363233375f30
+
+**05-07 QwenPaw Developer Day: HiClaw Overview & Using with QwenPaw**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643238333238393839375f323034353035363233375f39
+
+**04-23 QwenPaw Developer Day: Community Best Practice — QwenPaw for AI Short Drama Production**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643237313633343836315f323034353035363233375f30
+
+**04-21 QwenPaw Developer Day: Memory Proactive & Dream**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643236393039373133375f323034353035363233375f39
+
+**04-16 QwenPaw Developer Day: QwenPaw Skills**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643236343133303235385f323034353035363233375f30
+
+**04-14 QwenPaw Developer Day: QwenPaw Plugin Mechanism**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643236313434363337385f323034353035363233375f30
+
+**04-09 CoPaw Developer Day: AgentScope & CoPaw Memory & Context Management Upgrade**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643235363338303637335f323034353035363233375f30
+
+**04-07 CoPaw Developer Day: Open Tasks Update**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643235333834343132365f323034353035363233375f30
+
+**04-02 CoPaw Developer Day: CoPaw-Flash — Custom Small Model for CoPaw**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643235303333303132395f323034353035363233375f30
+
+**03-31 CoPaw Developer Day: CoPaw 1.0 Release**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643234373632373334315f323034353035363233375f30
+
+**03-26 CoPaw Developer Day: Multi-Agent**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643234323932343730315f323034353035363233375f30
+
+**03-24 CoPaw Developer Day: Security Mechanism Upgrade**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643234303532393639385f323034353035363233375f30
+
+**03-20 CoPaw Developer Day: v0.1.0 Release Update**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643233353633393630395f323034353035363233375f30
+
+**03-19 CoPaw Developer Day: CoPaw Memory & ReMe + Community Group Guide**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643233343536363335355f323034353035363233375f30
+
+**03-18 CoPaw Developer Day: Requirements Summary & Version Iteration**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643233333339353233355f323034353035363233375f30
+
+**03-17 CoPaw Developer Day: Unit Testing**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643233323231373438385f323034353035363233375f30
+
+**03-16 CoPaw Developer Day: Multi-Agent Workspace**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643233313136333431365f323034353035363233375f30
+
+**03-13 CoPaw Developer Day: CoPaw Security Mechanism**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643232383938333431365f323034353035363233375f30
+
+**03-11 CoPaw Developer Day: CoPaw Provider Refactor**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643232363639333732345f323034353035363233375f30
+
+**03-10 CoPaw Developer Day: CoPaw Release Update & Publishing Industry Developer Share**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643232353739353035315f323034353035363233375f30
+
+**03-09 CoPaw Developer Day: CoPaw Roadmap**
+
+Meeting link: https://shanji.dingtalk.com/app/transcribes/76327569643232343332383030385f323034353035363233375f39
