+++ website/public/blog/qwenpaw-developer-day-collection.zh.md
@@ -1,0 +1,148 @@
+---
+title: "QwenPaw 开发者日会合集"
+date: 2026-06-30
+author: QwenPaw Team
+tags: [QwenPaw, 日会合集]
+cover: /blog/qwenpaw-developer-day-collection-cover.png
+excerpt: "QwenPaw团队召开开发者日会，为每一位 QwenPaw 开发者与爱好者提供一份兼具理论深度与落地价值的完整技术交流档案。"
+---
+
+最近更新 2026 年 6 月 30 日
+
+---
+
+**06-30 QwenPaw 开发者日会：QwenPaw 2.0 Driver 模块详解**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643334363335313437375f323034353035363233375f30
+
+**06-25 QwenPaw 开发者日会：AgentScope X OceanBase 联合直播**
+
+会议链接：https://weixin.qq.com/sph/AO91dHFgqM
+
+**06-23 QwenPaw 开发者日会：QwenPaw 控制台更新以及用户消息队列**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643333373935343937395f323034353035363233375f30
+
+**06-16 QwenPaw 开发者日会：QwenPaw 插件分享 PawGit 和 TeamChat**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643333313236353132335f323034353035363233375f30
+
+**06-11 QwenPaw 开发者日会：QwenPaw 插件市场**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643332353835303430355f323034353035363233375f30
+
+**06-09 QwenPaw 开发者日会：QwenPaw Coding 模式**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643332323838383539345f323034353035363233375f30
+
+**06-04 QwenPaw 开发者日会：ReMe 升级**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643331373333363833335f323034353035363233375f30
+
+**06-02 QwenPaw 开发者日会：QwenPaw + Rokid Glass = ? 蚂蚁黑客松入围决赛项目「Third Eye」分享**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643331343533333032325f323034353035363233375f30
+
+**05-28 QwenPaw 开发者日会：AgentScope 2.0 详解**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643330383738303630325f323034353035363233375f30
+
+**05-26 QwenPaw 开发者日会：官方插件 CloudPaw 详解**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643330353931383433345f323034353035363233375f30
+
+**05-21 QwenPaw 开发者日会：定时任务与收件箱升级**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643330303635303130395f323034353035363233375f30
+
+**05-19 QwenPaw 开发者日会：最佳实践分享——基于 QwenPaw 打造个人工作站**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643239373239363837315f323034353035363233375f30
+
+**05-14 QwenPaw 开发者日会：QwenPaw 插件系统升级详解**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643239313839373636305f323034353035363233375f30
+
+**05-12 QwenPaw 开发者日会：QwenPaw Doctor 简介**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643238393133373436365f323034353035363233375f30
+
+**05-07 QwenPaw 开发者日会：HiClaw 简介及如何配合 QwenPaw 使用**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643238333238393839375f323034353035363233375f39
+
+**04-23 QwenPaw 开发者日会：社区资深开发者分享 QwenPaw 在 AI 短剧制作场景的最佳实践**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643237313633343836315f323034353035363233375f30
+
+**04-21 QwenPaw 开发者日会：Memory 的 Proactive 和 Dream**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643236393039373133375f323034353035363233375f39
+
+**04-16 QwenPaw 开发者日会：QwenPaw Skills**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643236343133303235385f323034353035363233375f30
+
+**04-14 QwenPaw 开发者日会：QwenPaw 插件机制**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643236313434363337385f323034353035363233375f30
+
+**04-09 CoPaw 开发者日会：AgentScope & CoPaw 记忆和上下文管理系统升级**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643235363338303637335f323034353035363233375f30
+
+**04-07 CoPaw 开发者日会：Open Tasks 更新**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643235333834343132365f323034353035363233375f30
+
+**04-02 CoPaw 开发者日会：专为 CoPaw 定制的小模型 CoPaw-Flash**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643235303333303132395f323034353035363233375f30
+
+**03-31 CoPaw 开发者日会：CoPaw 1.0 版本发布**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643234373632373334315f323034353035363233375f30
+
+**03-26 CoPaw 开发者日会：多智能体**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643234323932343730315f323034353035363233375f30
+
+**03-24 CoPaw 开发者日会：安全机制升级**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643234303532393639385f323034353035363233375f30
+
+**03-20 CoPaw 开发者日会：0.1.0 版本更新**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643233353633393630395f323034353035363233375f30
+
+**03-19 CoPaw 开发者日会：CoPaw Memory & ReMe 记忆系统解读 + 交流群分流**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643233343536363335355f323034353035363233375f30
+
+**03-18 CoPaw 开发者日会：需求汇总 & 版本迭代**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643233333339353233355f323034353035363233375f30
+
+**03-17 CoPaw 开发者日会：单测相关**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643233323231373438385f323034353035363233375f30
+
+**03-16 CoPaw 开发者日会：多智能体 Workspace**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643233313136333431365f323034353035363233375f30
+
+**03-13 CoPaw 开发者日会：CoPaw 安全机制**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643232383938333431365f323034353035363233375f30
+
+**03-11 CoPaw 开发者日会：CoPaw Provider 重构**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643232363639333732345f323034353035363233375f30
+
+**03-10 CoPaw 开发者日会：CoPaw 版本更新 & 出版行业开发者分享**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643232353739353035315f323034353035363233375f30
+
+**03-09 CoPaw 开发者日会：CoPaw Roadmap 介绍**
+
+会议链接：https://shanji.dingtalk.com/app/transcribes/76327569643232343332383030385f323034353035363233375f39
