+++ website/src/components/Icon/BolgsIcon.tsx
@@ -1,0 +1,28 @@
+import React from "react";
+import type { IconProps } from "./types";
+
+export const BolgsIcon: React.FC<IconProps> = ({
+  size = 18,
+  className = "",
+  color = "currentColor",
+}) => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      width={size}
+      height={size}
+      viewBox="0 0 16 16"
+      className={`inline-block transition-transform duration-300 ${className}`}
+    >
+      <path
+        d="M2.7876922186761472,2.2630730302755735Q2.6998536586761475,2.4672856302755735,2.6998536586761475,2.6902929402755738L2.6998536586761475,13.310292040275574Q2.6998536586761475,13.533303040275573,2.7876916456761474,13.737515040275573Q2.8720079086761476,13.933540040275574,3.0250213486761472,14.084002040275573Q3.1771546886761475,14.233600040275574,3.3742982786761475,14.315593040275573Q3.5779521486761476,14.400292040275573,3.7998536586761475,14.400292040275573L12.199854658676147,14.400292040275573Q12.421759358676148,14.400292040275573,12.625412658676147,14.315593040275573Q12.822558658676147,14.233600040275574,12.974691658676148,14.084000040275575Q13.127703658676147,13.933539040275575,13.212019658676148,13.737511040275574Q13.299853658676147,13.533302040275574,13.299853658676147,13.310292040275574L13.299853658676147,5.301780140275573Q13.299853658676147,5.080571740275573,13.214344658676147,4.876561740275574Q13.128834658676148,4.672551740275574,12.971108658676147,4.517453540275573L10.325527458676149,1.9159664802755738Q10.171226758676148,1.7642380402755737,9.970952258676148,1.6822657282755737Q9.770676858676147,1.6002930402755737,9.554274358676148,1.6002930402755737L3.7998537586761474,1.6002930402755737Q3.5779505386761477,1.6002930402755737,3.3742946986761475,1.6849964782755738Q3.1771522186761474,1.7669909302755737,3.0250196186761475,1.9165880702755738Q2.8720086686761475,2.0670488802755735,2.7876922186761472,2.2630730302755735ZM3.6998536586761475,13.310292040275574L3.6998536586761475,2.6902929402755738Q3.6998536586761475,2.6554790402755737,3.7261595586761476,2.6296116402755736Q3.755975158676147,2.6002929802755737,3.7998536586761475,2.6002929802755737L9.554274358676148,2.6002929802755737Q9.595204558676148,2.6002929802755737,9.624388958676146,2.6289911402755735L12.269968758676148,5.230476940275574Q12.299854058676148,5.259864940275573,12.299854058676148,5.301780140275573L12.299854058676148,13.310292040275574Q12.299854058676148,13.345113040275574,12.273551658676148,13.370978040275574Q12.243739858676147,13.400292040275573,12.199854658676147,13.400292040275573L3.7998536586761475,13.400292040275573Q3.7559717586761474,13.400292040275573,3.7261577586761474,13.370975040275574Q3.6998536586761475,13.345111040275574,3.6998536586761475,13.310292040275574ZM4.999853658676147,7.000292640275574L4.999853658676147,5.4002930402755736Q4.999853658676147,5.228623740275573,5.063044058676148,5.069007540275574Q5.131655658676147,4.895698640275574,5.2634577586761475,4.763896840275574Q5.395259658676148,4.632094940275573,5.568568658676147,4.563483640275574Q5.7281844586761475,4.5002931402755735,5.899853658676148,4.5002931402755735L9.899853958676147,4.5002931402755735Q10.071522958676148,4.5002931402755735,10.231138958676148,4.563483640275574Q10.404448258676148,4.632094940275573,10.536249858676147,4.763896840275574Q10.668051458676148,4.895698640275574,10.736662658676147,5.069007540275574Q10.799854058676148,5.228624440275574,10.799854058676148,5.4002930402755736L10.799854058676148,7.000292640275574Q10.799854058676148,7.171961140275574,10.736662658676147,7.331577640275574Q10.668051458676148,7.504886940275574,10.536249858676147,7.636688540275574Q10.404448258676148,7.768490640275574,10.231138958676148,7.837102740275574Q10.071522958676148,7.900293240275574,9.899853958676147,7.900293240275574L5.899853658676148,7.900293240275574Q5.728184658676147,7.900293240275574,5.568568658676147,7.837102740275574Q5.395259658676148,7.768490640275574,5.2634577586761475,7.636688540275574Q5.131655658676147,7.504886940275574,5.063044058676148,7.331577640275574Q4.999853658676147,7.1719616402755735,4.999853658676147,7.000292640275574ZM5.999853658676147,5.5002931402755735L5.999853658676147,6.900292740275574L9.799853558676148,6.900292740275574L9.799853558676148,5.5002931402755735L5.999853658676147,5.5002931402755735ZM10.499849058676148,9.020311740275574L5.499848858676147,9.020311740275574C5.234752158676148,9.020311740275574,5.019848858676148,9.235216040275574,5.019848858676148,9.500312240275573C5.019848858676148,9.765409340275573,5.234752158676148,9.980312240275573,5.499848858676147,9.980312240275573L10.499849058676148,9.980312240275573C10.764945758676147,9.980312240275573,10.979849558676147,9.765409340275573,10.979849558676147,9.500312240275573C10.979849558676147,9.235216040275574,10.764945758676147,9.020311740275574,10.499849058676148,9.020311740275574ZM8.499849058676148,10.920312740275573L5.499848858676147,10.920312740275573C5.234752158676148,10.920312740275573,5.019848858676148,11.135215640275574,5.019848858676148,11.400312340275574C5.019848858676148,11.665409040275573,5.234752158676148,11.880312040275573,5.499848858676147,11.880312040275573L8.499849058676148,11.880312040275573C8.764945758676149,11.880312040275573,8.979849058676148,11.665409040275573,8.979849058676148,11.400312340275574C8.979849058676148,11.135215640275574,8.764945758676149,10.920312740275573,8.499849058676148,10.920312740275573Z"
+        fillRule="evenodd"
+        fill={color}
+        fillOpacity="0.8799999952316284"
+      />
+    </svg>
+  );
+};
+
+export default BolgsIcon;
