+++ website/src/components/Icon/index.tsx
@@ -18,3 +18,4 @@ export { PathIcon } from "./PathIcon";
 export { WChatIcon } from "./WChatIcon";
 export { DouyinIcon } from "./DouyinIcon";
 export { DingTalkIcon } from "./DingTalkIcon";
+export { BolgsIcon } from "./BolgsIcon";
