+++ website/src/components/Nav.tsx
@@ -5,7 +5,7 @@ import { QwenpawMascot } from "./QwenpawMascot";
 import { useTranslation } from "react-i18next";
 import { useSiteLanguage } from "@/i18n/SiteLanguageContext";
 import { useSiteConfig } from "@/config-context";
-import { GitHubIcon, SelectTextIcon, NoteIcon } from "./Icon";
+import { GitHubIcon, BolgsIcon, NoteIcon } from "./Icon";
 
 const AGENTSCOPE_LOGO_SIZE = 22;
 
@@ -92,8 +92,12 @@ export function Nav() {
         </Link>
         <div className="nav-links hidden min-[641px]:flex min-[641px]:items-center min-[641px]:gap-6 lg:gap-8">
           <Link to={docsBase} className={navLinkOrangeClass}>
-            <SelectTextIcon />
+            <BookOpen size={18} strokeWidth={navIconStroke} aria-hidden />
             <span>{t("nav.docs")}</span>
+          </Link>
+          <Link to="/blog" className={navLinkOrangeClass}>
+            <BolgsIcon size={18} aria-hidden />
+            <span>{t("nav.blog")}</span>
           </Link>
           <a
             href="https://github.com/agentscope-ai/QwenPaw"
@@ -162,6 +166,13 @@ export function Nav() {
           onClick={() => setOpen(false)}
         >
           <BookOpen size={18} strokeWidth={navIconStroke} /> {t("nav.docs")}
+        </Link>
+        <Link
+          to="/blog"
+          className={navLinkOrangeClass}
+          onClick={() => setOpen(false)}
+        >
+          <BolgsIcon size={18} aria-hidden /> {t("nav.blog")}
         </Link>
         <a
           href="https://github.com/agentscope-ai/QwenPaw"
